Add files via upload
This code is for Round Robin Scheduling Algorithm. To run this code you have to save the RoundRobin.py file and run this. In the output window it will take the number of process than for each processes name, Burst time,Arrival Time and at last  quantum time . Than the output show you the calculation of each process's  and .Compute the average turnaround time and average waiting time.Provide a Gantt chart representing the scheduling order.